refactor(select): scaffold ui/select module and relocate value-chips
